backend: move airtable.js and twitter.js to api/ folder
